+++ sdk/auth-ui/src/ui/dialog.tsx
@@ -1,8 +1,6 @@
-"use client"
-
 import * as React from "react"
 import { Dialog as DialogPrimitive } from "@base-ui/react/dialog"
-import { cn } from "#authui/lib/utils"
+import { cn } from "cn"
 
 import { Button } from "#authui/ui/button"
 import { HugeiconsIcon } from "@hugeicons/react"