+++ sdk/auth-ui/src/ui/input-otp.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { cn } from "#authui/lib/utils"
+import { cn } from "cn"
 import { OTPInput, OTPInputContext } from "input-otp"
 import { HugeiconsIcon } from "@hugeicons/react"
 import { MinusSignIcon } from "@hugeicons/core-free-icons"