+++ sdk/auth-ui/src/ui/separator.tsx
@@ -1,5 +1,7 @@
+"use client"
+
 import { Separator as SeparatorPrimitive } from "@base-ui/react/separator"
-import { cn } from "#authui/lib/utils"
+import { cn } from "cn"
 
 function Separator({
   className,